Added Key class for managing keys (including mouse buttons and gamepad buttons).  Added InputClient for testing input module.  Fixed equality operator incorrectly made an assignment operator in GlobalObjectID.

+++ ArcanaEngine/ArcanaCore/KeyEvent.h
@@ -17,7 +17,7 @@ namespace Arcana
 			Released
 		};
 	
-		KeyEvent(Type event, uint64 keyCode, bool alt, bool control, bool shift, bool system) : Event(1)
+		KeyEvent(Type event, int32 keyCode, bool alt, bool control, bool shift, bool system) : Event(1)
 		{
 			getData().addInt("event", event);
 			getData().addInt("keyCode", keyCode);
